# Conflicts:
#	Project/Client/MeshRenderUI.cpp
#	Project/Client/TestLevel.cpp
#	Project/Engine/CMaterial.h
#	Project/Engine/CMeshRender.cpp
#	Project/Engine/CRenderComponent.h

+++ Project/Engine/CSetColorShader.cpp
@@ -19,8 +19,8 @@ void CSetColorShader::UpdateData()
 	m_OutTex->UpdateData_CS(0, false);
 
 	// 그룹 개수 계산
-	m_iGroupX = m_OutTex->Width() / m_iGroupPerThreadX;
-	m_iGroupY = m_OutTex->Height() / m_iGroupPerThreadY;
+	m_iGroupX = (UINT)m_OutTex->Width() / m_iGroupPerThreadX;
+	m_iGroupY = (UINT)m_OutTex->Height() / m_iGroupPerThreadY;
 	m_iGroupZ = 1;
 }
